requirements.txt
# OpenStack2NetBox requirements
pynetbox~=7.5.0
python-dotenv>=1.0.1
python-cinderclient~=9.5.0
python-keystoneclient~=5.4.0
python-neutronclient~=11.3.1
python-novaclient~=18.6.0
